senana/langchain/requirements.txt
requests>=2.28.0
pyyaml>=6.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-ollama>=0.0.2
